--- a/.gitignore
+++ b/.gitignore
@@ -17,3 +17,4 @@ vendor/
 go.sum
 
 oppai/logger
+e2e/