--- a/oppai/router.go
+++ b/oppai/router.go
@@ -10,6 +10,11 @@ type router struct {
 	handlers map[string]HandlerFunc
 }
 
+type oppaiHandlerCfg struct {
+	// ctx
+	OppaiCtx *Context
+}
+
 func NewRouter() *router {
 	return &router{
 		handlers: make(map[string]HandlerFunc),
@@ -75,30 +80,14 @@ func (r *router) getRoutes(method string) []*node {
 	return nodes
 }
 
-func (r *router) handle(ctx *Context) {
-	n, params := r.getRoute(ctx.Method, ctx.Path)
+func (r *router) handle(ctx *oppaiHandlerCfg) {
+	n, params := r.getRoute(ctx.OppaiCtx.Method, ctx.OppaiCtx.Path)
 	if n != nil {
-		ctx.Params = params
-		key := ctx.Method + "_" + n.pattern
-		r.handlers[key](ctx)
+		ctx.OppaiCtx.Params = params
+		key := ctx.OppaiCtx.Method + "_" + n.pattern
+		r.handlers[key](ctx.OppaiCtx)
 	} else {
-		ctx.String(http.StatusNotFound, "ERROR 404 NOT FOUND: %s\n", ctx.Path)
+		http.Error(ctx.OppaiCtx.Writer, "NOT FOUND", http.StatusNotFound)
 	}
 }
 
-// Only one * is allowed
-func parsePattern(pattern string) []string {
-	split := strings.Split(pattern, "/")
-
-	parts := make([]string, 0)
-	for _, item := range split {
-		if item != "" {
-			parts = append(parts, item)
-			if item[0] == '*' {
-				break
-			}
-		}
-	}
-
-	return parts
-}