--- a/oppai/router.go
+++ b/oppai/router.go
@@ -2,7 +2,6 @@ package oppai
 
 import (
 	"net/http"
-	"strings"
 )
 
 type router struct {
@@ -11,7 +10,6 @@ type router struct {
 }
 
 type oppaiHandlerCfg struct {
-	// ctx
 	OppaiCtx *Context
 }
 
@@ -22,14 +20,11 @@ func NewRouter() *router {
 	}
 }
 
-func (r *router) AddRoutes(method string, pattern string, handler HandlerFunc) {
-
+func (r *router) AddRoutes(method, pattern string, handler HandlerFunc) {
 	parts := parsePattern(pattern)
-
 	key := method + "_" + pattern
-	_, ok := r.root[method]
 
-	if !ok {
+	if _, ok := r.root[method]; !ok {
 		r.root[method] = &node{}
 	}
 
@@ -37,7 +32,7 @@ func (r *router) AddRoutes(method string, pattern string, handler HandlerFunc) {
 	r.handlers[key] = handler
 }
 
-func (r *router) getRoute(method string, path string) (*node, map[string]string) {
+func (r *router) getRoute(method, path string) (*node, map[string]string) {
 	searchParts := parsePattern(path)
 	params := make(map[string]string)
 	root, ok := r.root[method]
@@ -47,37 +42,19 @@ func (r *router) getRoute(method string, path string) (*node, map[string]string)
 	}
 
 	n := root.search(searchParts, 0)
-
 	if n != nil {
 		parts := parsePattern(n.pattern)
 		for i, part := range parts {
 			if part[0] == ':' {
 				params[part[1:]] = searchParts[i]
-			}
-
-			if part[0] == '*' && len(part) > 0 {
-				params[part[1:]] = strings.Join(searchParts[i:], "/")
+			} else if part[0] == '*' {
+				params[part[1:]] = joinParts(searchParts[i:])
 				break
 			}
 		}
-
 		return n, params
 	}
-
 	return nil, nil
-}
-
-func (r *router) getRoutes(method string) []*node {
-	root, ok := r.root[method]
-
-	if !ok {
-		return nil
-	}
-
-	nodes := make([]*node, 0)
-	root.travel(&nodes)
-
-	return nodes
 }
 
 func (r *router) handle(ctx *oppaiHandlerCfg) {
@@ -91,3 +68,20 @@ func (r *router) handle(ctx *oppaiHandlerCfg) {
 	}
 }
 
+func joinParts(parts []string) string {
+	length := 0
+	for _, part := range parts {
+		length += len(part) + 1
+	}
+	joined := make([]byte, length-1)
+	pos := 0
+	for _, part := range parts {
+		copy(joined[pos:], part)
+		pos += len(part)
+		if pos < length-1 {
+			joined[pos] = '/'
+			pos++
+		}
+	}
+	return string(joined)
+}