- improve node & router
# Before Improvement
`fanchann@fanchann-20205 ~ $ plow -m GET -d 2m -c 4 http://localhost:9000/say/ahihahia`
```sh
Benchmarking http://localhost:9000/say/ahihahia for 2m0s using 4 connection(s).
@ Real-time charts is listening on http://[::]:18888

Summary:
  Elapsed       2m0s
  Count      1258756
    2xx      1258756
  RPS      10489.421
  Reads    1.571MB/s
  Writes   0.700MB/s

Statistics    Min      Mean    StdDev     Max   
  Latency    48µs     376µs    1.382ms  90.934ms
  RPS       6228.26  10489.42  1409.34  17417.06

Latency Percentile:
  P50     P75    P90    P95     P99     P99.9     P99.99 
  173µs  269µs  455µs  799µs  5.253ms  20.068ms  42.326ms

Latency Histogram:
  320µs     1245293  98.93%
  3.391ms     10457   0.83%
  9.896ms      2245   0.18%
  20.007ms      605   0.05%
  33.343ms      113   0.01%
  44.886ms       27   0.00%
  54.973ms       15   0.00%
  90.934ms        1   0.00%
```
# After
`fanchann@fanchann-20205 ~ $ plow -m GET -d 2m -c 4 http://localhost:9000/say/ahihahia`
```sh
Benchmarking http://localhost:9000/say/ahihahia for 2m0s using 4 connection(s).
@ Real-time charts is listening on http://[::]:18888

Summary:
  Elapsed       2m0s
  Count      1254171
    2xx      1254171
  RPS      10451.245
  Reads    1.565MB/s
  Writes   0.698MB/s

Statistics    Min      Mean    StdDev     Max   
  Latency    48µs     378µs    1.424ms  97.892ms
  RPS       4172.64  10450.83  2148.32  17408.32

Latency Percentile:
  P50     P75    P90    P95     P99     P99.9     P99.99 
  170µs  267µs  455µs  831µs  5.174ms  20.234ms  45.097ms

Latency Histogram:
  265µs     1147069  91.46%
  733µs       89659   7.15%
  3.639ms     13494   1.08%
  10.677ms     3065   0.24%
  21.173ms      680   0.05%
  38.003ms      163   0.01%
  50.558ms       34   0.00%
  61.718ms        7   0.00%
```
# improvement #2
`fanchann@fanchann-20205 ~ $ plow -m GET -d 2m -c 4 http://localhost:9000/say/ahihahia`
```sh
Benchmarking http://localhost:9000/say/ahihahia for 2m0s using 4 connection(s).
@ Real-time charts is listening on http://[::]:18888

Summary:
  Elapsed       2m0s
  Count      1333026
    2xx      1333026
  RPS      11108.448
  Reads    1.663MB/s
  Writes   0.742MB/s

Statistics    Min     Mean    StdDev      Max   
  Latency    49µs     355µs   1.287ms  118.218ms
  RPS       6665.81  11108.4  1364.26  17969.18 

Latency Percentile:
  P50     P75    P90    P95     P99     P99.9     P99.99 
  170µs  262µs  443µs  768µs  4.404ms  18.927ms  41.653ms

Latency Histogram:
  281µs     1278259  95.89%
  992µs       45019   3.38%
  4.853ms      7493   0.56%
  11.284ms     1619   0.12%
  20.186ms      492   0.04%
  31.611ms      101   0.01%
  46.812ms       39   0.00%
  51.846ms        4   0.00%
```
# improvement #3
`fanchann@fanchann-20205 ~ $ plow -m GET -d 2m -c 4 http://localhost:9000/say/ahihahia`
```sh
Benchmarking http://localhost:9000/say/ahihahia for 2m0s using 4 connection(s).
@ Real-time charts is listening on http://[::]:18888

Summary:
  Elapsed       2m0s
  Count      1349975
    2xx      1349975
  RPS      11249.512
  Reads    1.684MB/s
  Writes   0.751MB/s

Statistics    Min      Mean    StdDev     Max   
  Latency    46µs     351µs    1.282ms  75.225ms
  RPS       5605.58  11249.06  2145.1   19487.46

Latency Percentile:
  P50     P75    P90    P95     P99     P99.9     P99.99 
  165µs  254µs  427µs  738µs  4.519ms  18.459ms  38.663ms

Latency Histogram:
  299µs     1327641  98.35%
  2.021ms     19701   1.46%
  10.545ms     2093   0.16%
  23.964ms      423   0.03%
  36.488ms      103   0.01%
  48.235ms       11   0.00%
  53.33ms         2   0.00%
  69.713ms        1   0.00%
```

# improvement #4
`fanchann@fanchann-20205 ~ $ plow -m GET -d 2m -c 4 http://localhost:9000/say/ahihahia`
```sh
Benchmarking http://localhost:9000/say/ahihahia for 2m0s using 4 connection(s).
@ Real-time charts is listening on http://[::]:18888

Summary:
  Elapsed       2m0s
  Count      1380965
    2xx      1380965
  RPS      11507.982
  Reads    1.723MB/s
  Writes   0.768MB/s

Statistics    Min      Mean    StdDev     Max   
  Latency    48µs     343µs    1.249ms  83.14ms 
  RPS       6777.71  11512.34  2203.7   19589.29

Latency Percentile:
  P50     P75    P90    P95     P99     P99.9     P99.99 
  165µs  256µs  429µs  729µs  4.616ms  18.307ms  41.357ms

Latency Histogram:
  280µs     1312484  95.04%
  776µs       57619   4.17%
  3.757ms      8659   0.63%
  10.396ms     1735   0.13%
  19.854ms      370   0.03%
  34.423ms       80   0.01%
  49.676ms       16   0.00%
  68.203ms        2   0.00%
```
- feature added:
## cookie (set & get)
## Bind
## Render html File
## Get IP address
## Simplify Configuration
and more